Change API path of app-wise token revocation API
