+++ components/org.wso2.carbon.identity.api.user.authorized.apps/org.wso2.carbon.identity.rest.api.user.authorized.apps.v2/src/main/resources/authorizedApps.yaml
@@ -233,7 +233,7 @@ paths:
         '500':
           $ref: '#/components/responses/ServerError'
 
-  '/applications/{application-id}/tokens':
+  '/authorized-apps/{application-id}/tokens':
     delete:
       tags:
         - admin
